haiku-generator/requirements.txt
anthropic>=0.18.0
python-dotenv>=1.0.0
numpy>=1.24.0
pronouncing>=0.2.0
openai>=1.0.0
Pillow>=10.0.0
